+++ Demo/Demo/Resources/Image+Demo.swift
@@ -10,12 +10,20 @@ import SwiftUI
 
 extension Image {
     
-    static var alert: Image { Image(systemName: "exclamationmark.triangle") }
-    static var checkmark: Image { Image(systemName: "checkmark") }
-    static var clear: Image { Image(systemName: "xmark.circle") }
-    static var dismiss: Image { Image(systemName: "xmark") }
-    static var navigationArrow: Image { Image(systemName: "chevron.right") }
-    static var text: Image { Image(systemName: "abc") }
-    static var type: Image { Image(systemName: "square.and.pencil") }
-    static var settings: Image { Image(systemName: "gearshape") }
+    static let alert = Image.symbol("exclamationmark.triangle")
+    static let checkmark = Image.symbol("checkmark")
+    static let clear = Image.symbol("xmark.circle")
+    static let dismiss = Image.symbol("xmark")
+    static let text = Image.symbol("abc")
+    static let type = Image.symbol("square.and.pencil")
+    static let safari = Image.symbol("safari")
+    static let settings = Image.symbol("gearshape")
+    
+    static func file(_ name: String) -> Image {
+        Image(name)
+    }
+    
+    static func symbol(_ name: String) -> Image {
+        Image(systemName: name)
+    }
 }